+++ lib/src/televerse/models/models.dart
@@ -13,3 +13,4 @@ part 'webhook_exception.dart';
 part 'http_exception.dart';
 part 'long_polling_exception.dart';
 part 'televerse_exception.dart';
+part 'mention.dart';