+++ lib/televerse.dart
@@ -35,6 +35,7 @@ export 'src/types/types.dart';
 export 'src/televerse/event/event.dart';
 export 'src/televerse/filters/filters.dart';
 export 'src/televerse/fetch/fetch.dart';
+export 'src/televerse/extensions/extensions.dart';
 
 part 'src/televerse/televerse.dart';
 part 'src/utils/date.dart';